+++ system_report.sh
@@ -61,6 +61,16 @@ which adb
 echo
 echo " * adb version:"
 adb version
+echo
+echo " * SDK packages:"
+if [[ ! -z "${BITRISE_DOCKER_REV_NUMBER_ANDROID_NDK_LTS}" ]] ; then
+    echo " (!) Version check not available on this Stack / in this image"
+else
+    ver_line="$(grep ^Pkg.Revision ${ANDROID_HOME}/tools/source.properties | cut -d= -f 2)" ; echo "* SDK Tools version: $ver_line"
+    ver_line="$(grep ^Pkg.Revision ${ANDROID_HOME}/platform-tools/source.properties | cut -d= -f 2)" ; echo "* Platform Tools version: $ver_line"
+    ver_line="$(grep ^Pkg.Revision ${ANDROID_HOME}/emulator/source.properties | cut -d= -f 2)" ; echo "* Emulator version: $ver_line"
+fi
+
 echo "========================================"
 echo